+++ src/jsonyx/_speedups.c
@@ -1094,6 +1094,11 @@ scan_once_unicode(PyScannerObject *s, PyObject *memo, PyObject *pyfilename, PyOb
                 return NULL;
             res = _parse_object_unicode(s, memo, pyfilename, pystr, idx + 1, next_idx_ptr);
             _Py_LeaveRecursiveCall();
+            if (PyErr_ExceptionMatches(PyExc_RecursionError)) {
+                PyErr_Clear();
+                raise_errmsg("Object is too deeply nested", pyfilename, pystr, idx, 0);
+                return NULL;
+            }
             return res;
         case '[':
             /* array */
@@ -1102,6 +1107,11 @@ scan_once_unicode(PyScannerObject *s, PyObject *memo, PyObject *pyfilename, PyOb
                 return NULL;
             res = _parse_array_unicode(s, memo, pyfilename, pystr, idx + 1, next_idx_ptr);
             _Py_LeaveRecursiveCall();
+            if (PyErr_ExceptionMatches(PyExc_RecursionError)) {
+                PyErr_Clear();
+                raise_errmsg("Array is too deeply nested", pyfilename, pystr, idx, 0);
+                return NULL;
+            }
             return res;
         case 'n':
             /* null */