+++ docs/source/changelog.rst
@@ -25,6 +25,7 @@ jsonyx 2.0.0 (unreleased)
 - Added :func:`jsonyx.paste_values`
 - Added :func:`jsonyx.Manipulator`
 - Changed error for big integers to :exc:`jsonyx.JSONSyntaxError`
+- Changed error for deep nesting to :exc:`jsonyx.JSONSyntaxError`
 - Fixed line comment detection
 - Fixed spelling of "commas" in error messages
 - Fixed ``end_offset`` of :exc:`jsonyx.JSONSyntaxError`